requirements.txt
python-dotenv==0.21.1
openai==1.43.0
sentence-transformers==3.0.1
protobuf==3.20.*
pydantic==2.8.2
docstring_parser==0.16
faiss-cpu==1.8.0
numpy==1.26.4
matplotlib==3.9.2